Updates SpaceFieldRouter adapter (referent to Explosion class implementation)
This commit implements the sending of the right explosion to the
front-end via the JSON file. Note that this class that triggers the
explosion, so the definition of a triggered explosion (one that should
not be rendered again) is made by this endpoint (we need to further
evaluate this approach).